+++ R/knn.R
@@ -103,6 +103,17 @@ regression <- function(model, example) {
     prediction <- unname(colMeans(values))
   } else if (model$cf == "median") {
     prediction <- apply(values, 2, stats::median)
+  } else if (model$cf == "weighted") {
+    if (distances[o[1]] == 0) {
+      prediction <- unname(values[1, ])
+    } else {
+      reciprocal_d <- 1 / distances[o[1:model$k]]
+      prediction <- numeric(ncol(model$example$targets))
+      for (k in seq(model$k)) {
+        prediction <- prediction + values[k, ] * reciprocal_d[k]
+      }
+      prediction <- prediction / sum(reciprocal_d)
+    }
   }
   list(
     prediction = prediction,