+++ NEWS.md
@@ -1,3 +1,4 @@
 ## tsfknn 0.1.1
 
 * Fix calculation of how many KNN examples has the model in knn_forecasting
+* Weighted combination of the targets of nearest neighbors is implemented